refactor: move shared helpers out of component modules
