+++ apps/docs/src/components/landing/app-mock/tour.tsx
@@ -11,7 +11,8 @@ import * as m from 'motion/react-m';
 import { type ReactNode, useEffect, useState } from 'react';
 import { EASE } from '../motion';
 import { AppFrame } from './app-frame';
-import { Cursor, type Point, PRESS, travelTime } from './cursor';
+import { Cursor } from './cursor';
+import { type Point, PRESS, travelTime } from './cursor-motion';
 import { EventScene } from './event-scene';
 import { MockAgents } from './mock-agents';
 import { MockIssueDetail } from './mock-issue-detail';
