+++ apps/webview-ui/src/shared/ui/components/metric-delta.tsx
@@ -1,26 +1,8 @@
 import type { MetricDelta } from '@rustrak/client';
 import { ArrowDownRight, ArrowRight, ArrowUpRight } from 'lucide-react';
 import { formatPercentChange, percentChange } from '@/shared/lib/chart-format';
+import { deltaTone, type Polarity } from '@/shared/lib/metric-tone';
 import { cn } from '@/shared/lib/utils';
-
-/**
- * Whether a rise in this metric is good or bad news.
- *
- * Callers must say: more events is bad, more resolved issues is good. Colour
- * the arrow by direction alone and it tells the reader the opposite of the
- * truth half the time.
- */
-export type Polarity = 'up-is-bad' | 'up-is-good';
-
-export function deltaTone(change: number, polarity: Polarity): string {
-  if (change === 0) {
-    return 'text-muted-foreground';
-  }
-  const isGood = polarity === 'up-is-good' ? change > 0 : change < 0;
-  return isGood
-    ? 'text-green-600 dark:text-green-400'
-    : 'text-red-600 dark:text-red-400';
-}
 
 interface MetricDeltaTextProps {
   metric: MetricDelta;
